requirements.txt
docling>=2.10.0
phidata>=2.6.5
pandas>=2.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
python-dotenv>=1.0.0
